Less ad-hoc function calls, type marks and casts
When specifying function calls, type marks, and type casts in VHDL
it is now possible to use functions which make the codegen more
readable, e.g.

  functionCall "fn" [arg_0, arg_1, ..., arg_n]
  typeMark "type" expr
  typeCast "type" expr

A bonus of this is that it makes it easier to differentiate between
marks and casts when reading (as the only difference is the inclusion
of the single quote character).