Return "" for irrelevant comments in kloonbot
Attempting to exit from the parse_comment function in kloonbot
only exits the subshell, meaning exiting 0 would set the result to
the echoed message and then use that message as a command. Now
kloonbot will return the empty string, and then check if the parsed
comment is empty before continuing.