Qualify unsafeSynchronizer in Clash.Tutorial
When building tags in CI, -multiple-hidden is used everywhere (in
contrast to CI jobs on branches where it is on by default everywhere
except 8.6.5). This means the haddock job fails, as unqualified
references to unsafeSynchronizer in Clash.Tutorial will not refer
to anything in scope (as it is only exported from the explicit
prelude when multiple hidden is off).

(cherry picked from commit 05949cdf529b573db03a97863c8682a5352bf0a8)