Add missing warnings to Integer/Natural prims
The dynamic-size warning on primitives using Integer / Natural is
not consistently included in the .primtives files. This is down to
a mix of legacy (the warning existed after integer prims) and new
primitives being added and not having the warning (either by
forgetting or the primtives around them not having the warning).