Improve documentation on hidden arguments
This is a partial backport of #1849.

Documentation in Clash.Signal is corrected and extended. Parts of the
existing documentation were not rendered before this commit due to
incorrect interaction between CPP #ifdef and Haddock syntax.

A developer note [Going from WithSingleDomain to WithSpecificDomain] is
added in the source, explaining the code.