Fixes on Type related functions
Functions like piResultTys now only throw a +debug error for type
mismtaches when the coreviews of the types are not alpha-equivalent.
The partial evaluator can encounter some cases where an argument
unexpectedly has / does not have a Signal type.

coreView1 now looks through annotations on types, as these are
somewhat meta, and do not change the underlying type representation.