New unsafeSimSynchronizer, use in outputVerifier
Warn when `Clash.Explicit.Testbench.outputVerifier` is used in HDL which
is not marked as being a test bench and is given different clock domains
as `testDom` and `circuitDom`. If `testDom` and `circuitDom` refer to
the same domain, the warning is not emitted.

To this purpose, a new primitive `unsafeSimSynchronizer` is added to
`Clash.Explicit.Testbench`, and it is used if a clock domain crossing is
needed in `outputVerifier`.

The code for `outputVerifier` is also cleaned up. It now declares it
needs a non-empty list of samples to verify, since it never worked with
an empty list. Common code is split off and shared rather than
duplicated, and it now uses `Bounded` and `SaturatingNum` to work with
the `Index` iterating the samples instead of writing out the same
functionality.

Documentation for `biTbClockGen` is adjusted after discussion about
unclear parenthetical in original doc.

Rationale: `outputVerifier` does clock domain crossing in a manner that
is not suitable for actual hardware but is efficient for HDL simulation.
The code was not intended to be used outside of a test bench context,
but nonetheless it could be synthesized as part of hardware. If a
user does this, we should warn them that the resulting hardware is
improper.

The user can then add their own clock domain crossing in the circuit if
they would need this functionality. An imaginable use of
`outputVerifier` in a circuit is in combination with an ILA standing in
for the `assert` behavior.