Assorted doc fixes (#1941)
The documentation on the HaskellPrelude imports was missing in 2 of 4
places it was supposed to be. Rewritten to explain rather than
exhaustively list, also avoiding the need to update across all files
(Clash.Prelude.Safe was out-of-date already).

Added the `hasBlackBox` annotation to the example in the tutorial
because we should encourage our users to add that annotation.
Reformatted the import list in the example and added an import for
`hasBlackBox`.

In the example for `Clash.Explicit.Testbench.tbClockGen`, a signal
called `done'` (note prime) was defined as the logical inverse of the
`done` signal. Renamed it to `notDone` to avoid confusion.

(cherry picked from commit aabf61cd1f9b2a4851b11a8743e74586d5dd5a47)