Suggestion: store memory as literal string
When you want to generate the contents of a memory, rather than read it
in from a pre-generated file, one could use Template Haskell and
`qRunIO` to create it on the fly while compiling or loading into the
interpreter. Our documentation refers to this as something
["for those of us who like to live on the edge"](https://hackage.haskell.org/package/clash-prelude-1.5.0/candidate/docs/Clash-Sized-Fixed.html#creatingdatafiles).
It is brittle for many reasons.

Additionally, our `Vec` has performance issues when they get long, for a
pretty small value of "long". And lists don't work either.

So here is another idea. Using Template Haskell, we convert a list of
`BitPack`able values to the string contents of a file such as `romFile`
consumes. We don't write it to a file, though, but store it as a literal
string in our Haskell module. `romString` then is the variant of
`romFile` that uses this string while simulating in Haskell, and the
black box definition for `romString` uses the include mechanism to
write a file called `memXYZ.bin` next to the generated HDL, whereupon
the mechanism in HDL is identical to `romFile`'s mechanism.

When I create a `MemString` that holds 4 MiB of 32-bit data, the
resulting data file is 33 MiB (ASCII binary digits and newline), and so
is the object file for the Haskell module containing that `MemString`.
As can be seen from the `BlackBoxContext`, the `memStringContents` is
stored as `unpackCString# "literal"#`, and it is an efficient storage
method (well, apart from the binary digits bit of it, which we might
improve upon later by implementing hexadecimal digits for memory files).

If people like this idea, I can work it out further, this is a first
draft to gauge enthousiasm.