Move undefinedTy and unsafeCoerceTy to Util
Moving these types out of Clash.Core.Type means there is no longer
a circular import between Clash.Core.TysPrim and Clash.Core.Type.
This means we can access the whole of Clash.Core.Type without
needing to expose anything more in .hs-boot files.

We need this because the equality primitive type constructor (~#)
has a polymoprhic type which we need to manually reconstruct in
the TysPrim module, and currently cannot because not enough things
are available when importing boot modules.