Add eqPrimTyCon (~#) to Clash.Core.TysPrim
The primitive type equality type is the basis of all typical type
equality types you see in GHC (~ and ~~ are special within GHC but
defined as though they are subclasses of a ~# b). As such this is
the equality that often appears in GHC / Clash core when coercions
are used in code (e.g. GADT pattern matches).