Fix example topEntities in doc on hidden args
When I "fixed" these in #1849 I "fixed" them worse.

`resetSynchronizer` ignores its `Enable` argument, so it doesn't need
one passed in either; it would accomplish nothing and cause confusion.
And since the top entity contains its own PLL, any signal coming from
outside (the `ena` argument) will almost certainly not be synchronized
to the clock coming out of the PLL and should not be passed to `leds`.
So we should get rid of the `Enable` argument.