Add extra parens to rotate blackbox templates
When we have a blackbox template which wraps values to size, e.g.

```
to_integer(~ARG[1]
  -- pragma translate_off
  mod 32
  -- pragma translate_on
  );
```

this becomes invalid with bare literals if the argument expands to
something which has a unary minus at the start. Without bare literals
this would have rendered with parentheses around the entire argument,
which are now included in the template to ensure correct behaviour.