Keep recursive and non-recursive let in Clash Core
In GHC core, the difference between recursive and non-recursive
binders is preserved. In Clash, this difference is not preserved
in let expressions / the bindings map, which leads to sometimes
needing to calculate the connected components of the bindings.

However, since Letrec is used a lot (i.e. in normalization) it is
painful to simply differentiate everywhere now. As a compromise,
this is kept as a pattern synyonm so exsting code should work as
it did before. In the future we can migrate these cases until only
`Let NonRec{}` and `Let Rec{}` are used in code. Until this time,
the assumption is made that `NonRec` is _always_ non-recursive, but
`Rec` holds recursive and _potentially_ recursive bindings.

The parts of the compiler where it is easier to change the definitions
now have been changed.