Use SHA256 for cache invalidation
Before this commit, we used `hashable`s `hash`. For the purposes of
cache invalidation, it is very important that if `hash a == hash b`,
the chance that `a /= b` is vanishingly small. It is not clear that
`hashable` satisfies this requirement, whereas a cryptographically
secure hash such as SHA-256 clearly does.