clash-lib: Add enableToBit (#1984)
* clash-lib: Add enableToBit

Enable used to be a Bool in the DSL, so we could use `boolToBit`.
However it now has its own type in the DSL (Enable domainName), so we
need a new conversion function in order to convert it to a Bool.
This should work for all backends as they all currently normalize Enable
to a Bool.

* Add changelog entry

* Add copyright notice

(cherry picked from commit d059ab44c597922335d34d85dc7a031ec0d7de1b)