Look at Attributes and Ticks in Alpha Equality
When comparing types and terms for alpha equivalence, Clash used
a different approach for both with the common idea of ignoring
extra-syntactic elements. However, with the change to ordered maps
this results in collisions where previously there were none (as
they are used when hashing).

When specialising this is a problem, we want to ignore all the
extra-syntactic data on a type or term when comparing so we get
more cache hits in the specialisation cache. To preserve this, all
spec args in the cache keys are run through a new stripAllTicks
function which removes all ticks from a term. This same function
is used in recToLetRec to ensure the equality check still passes
as expected.