Change RewriteMonad to use RWST from `transformers`
The RewriteMonad as implemented was the same as the CPS version
of the transformer from transformers-0.5.6.0 onwards. Since we now
also want a monad based in IO so we can use `async`, it makes more
sense to use the transformers implementation (preferring the more
space efficient CPS version if available) instead of rewriting all
instances to use IO.

This commit also removes the MonadFail instance for RewriteMonad.
It was only used in one place, where a potentially failing pattern
match was used. This is replaced with a call to `either` which throws
a more meaningful error message on failures.