blockRam(File) primitive: fix undefined enable (#2054)
The Haskell models of `blockRam#` and `blockRamFile#` treated an _undefined_
write enable as an asserted enable. But an _undefined_ value in Haskell can
correspond to any value whatsoever in HDL, so HDL simulation might or might not
write. With this commit, the `XException` of the write enable is written as the
value in the RAM, since it could have either been written to or not been
written to. On the next read of that address, it will return the `XException`.

This issue did not propagate to any other `blockRam` variants, the bug solely
manifested when using the `blockRam#` and `blockRamFile#` primitives directly.
All the other variants built upon those primitives always have their write
address undefined whenever the write enable is undefined, and that case was
properly handled by the primitive.

The issue is identical to one of the issues in PR #2006 and PR #2031,
for different memory primitives.