Use new API for normalization
The new ClashEnv type is now used in the RewriteEnv, and any common
data between the ClashEnv and the RewriteEnv / NormalizeState is
removed and the references updated.

For minimal disruption, the old lenses generated for these are
reimplemented as top-level definitions of `Getter RewriteEnv a` in
Clahs.Rewrite.Types. The setters that were generated previously are
not needed, as no field in the ClashEnv is updated in rewriting.