Use `CACHE_FALLBACK_KEY` on GitLab CI (#2067)
We've been seeing a lot of dependency rebuilds, likely due to parallel
CI runs overwriting each other's cache. Using a branch-specific cache,
plus a fallback on master hopefully mitigates this.

I've also taken the opportunity to remove the `cabal-store` part of the
cache key, as it is used to cache more than just Cabal.