Efficient constants for initializing memories (#2041)
Define a new data type, `MemBlob`, and provide new memory primitive
variants that can use a `MemBlob` for their initial contents.

Currently, we provide two ways of initializing memories: with a `Vec`
and with a file. In some cases, `Vec` can be prohibitively inefficient;
depending on how the `Vec` is constructed, this can already happen with
a few tens of memory locations. Files allow for very large and arbitrary
contents, but generating the contents of the file in the same
compilation run as generating the HDL is very brittle and is a great way
to shoot oneself in the foot. This is why our current documentation
refers to this as "living on the edge".

This new mechanism dubbed `MemBlob` provides the user with Template
Haskell constructing functions which store the data alongside the
compiled Haskell code with very little overhead (max 7%, often no
overhead at all). The Template Haskell functions are the only way of
constructing a concrete `MemBlob`.

The `MemBlob` is a new Clash netlist type, and the template language for
primitives can refer to such an argument with ~CONST for its value, and
~LENGTH and others for its type information, just like with a normal
`Vec`. Primitives for the memory functions are very similar to the
existing primitives, but can even be slightly simpler because the data
is always of type `BitVector m`.

A basic functionality test for `asyncRom` is also included as I had
written the code anyway to debug a silly issue with `asyncRomBlob`.