tdpRAM: accurately model concurrent edges
Previously the fast port would run ahead and the slow port would
catch up. But especially for writes to the RAM, this is not
correct. When the fast and slow port are active concurrently,
both writes should commit, so that in the next cycle of the fast
port we can already read what's been written by the slow port.

Remember, on an actual multi-ported SRAM, both ports have the
same timing specifications. The "slow" port only indicates
that its clock period is longer than the clock of the "fast"
port.

Also, previously, when a write on the fast port occured on
the same address as a read on the slow port at any point
for the duration of a slow period, the output of the slow
port would be made invalid. However, actual multi-ported SRAMs
have output latches which are only transparent for a
short duration after read strobe of the RAM array.

This patch has therefor changes the read/write conflict behavior.
Outputs are now only invalidated due to read/write conflicts
when the ports are active concurrently; i.e. we are now
more optimistic with regards to read/write conflicts. The old
model was just too pessimistic.

In addition, this patch adds write/write conflict behavior when
both ports are writing concurrently.

(cherry picked from commit 212eb9a2a9a4172b3f4d15f651e3a99b477b47db)