Return boxed type in unpack{Float,Double}#
In the primitives unpackFloat# and unpackDouble#, the result from
the primitive evaluator was an unboxed literal (i.e. Float#) when
the actual result of the function is a boxed literal (i.e. Float).
This led to #2097 having core that looked like

```
case 1.0# of
  F# x -> ...
```

which will rightly never match. When the primitives instead return
the correct type we get a case expression

```
case F# 1.0# of
  F# x -> ...
```

where the caseCon transformation can fire.

(cherry picked from commit 31c25df1ba782fefdb3b158921cc41b3026b186c)