Fix missing libraries with snap package
When building the snap, snapcraft is filtering out several shared library
packages (e.g., libgmp10) because they are included in the core20 snap.  This
causes a problem when the -dev versions of these packages are used (e.g.,
libgmp-dev) where the clash snap will have broken symbolic links to the
shared library packages.  Work around this issue by staging all the relevant
library packages so they are included in the snap as well.

Fixes #2124.