Use Int instead of Int# in toEnum / fromEnum prims (#2127)
* Add test for #2126

* Use Int instead of Int# in toEnum / fromEnum prims

In the primitives for conversion to / from Enum introduced in 0613f21,
the definitions are converting to / from Int# instead of Int as the
type of the functions demands. This is not correct, and means that
conversion to the enum type will never fire (as the argument is always
a boxed int instead of an unboxed int), and the conversion from the
enum type to Int will result in an ill-typed expression being output
which can lead to

    case 1# of
      I# x -> ...

which are obviously bogus.

(cherry picked from commit a85df82b6a9454f36014f4e5a65277a49c866df6)