Fix and document Altera PLL's (#2136)
Commit d325557750 introduced an off-by-one error in the number of clocks
to generate for `alteraPll`.

Additionally, the internal `locked` signal for `alteraPll` is changed to
be of type `Bit` like in `altpll`. The borrowing of the type of the
`Reset` input probably came about out of convenience back when resets
were still a simple type so it did not matter. Now it looked odd but
worked fine nonetheless.

Also adjusted the type in `CondAssignment` in `altpll`, which showed the
same pattern.