Use blocking / non-blocking terminology
When referring to signals, we used Wire or Reg, which is (System)Verilog
specific terminology. When using VHDL, many times we render Reg as
a signal instead of variable, which is closer to the VHDL analogue.

To be more truthful in netlist, this is changed to blocking and
non-blocking, referring to how values are assigned. The VHDL should
be changed so that when Blocking is used, it makes a variable, e.g.

    block is
        variable foo : type;
    begin
        ...
    end block;