Add Netlist Declaration AST
In the revised netlist AST types, we want to be able to properly
differentiate between statements and declarations (which the old
netlist AST didn't). A new module is created for declarations
which currently includes

    * signal declarations
    * type declarations
    * constant declarations

as these are available in all HDL backends. Declarations can only
be included in statements by using blocks, i.e.

    * a concurrent block, which works like `block` in VHDL
    * a sequential block, which works like `process` in VHDL

When components are defined, the architecture of a component will
also be a block over a concurrent statement, which contains
declarations for the `architecture`.