Gitlab CI: automatically invalidate caches when docker image updates
As seen in #2201, updates to the docker images can cause stuff in the
cache to be broken.
By putting the docker image into the cache key we always start clean
whenever we update the docker images.

The `-3` is the cache bust suffix that gitlab automatically adds,
except to the fallback key. See:
https://docs.gitlab.com/ee/ci/caching/#clear-the-cache-manually
https://docs.gitlab.com/ee/ci/caching/#cache-key-names