Make newer haddock happy
The haddock shipped with GHC 9 was generating warnings about not being
able to link to `Tracked` and `Naming` because those types aren't being
exported.

But the docs making those references are on things which weren't
exported either.

The doc for the internal constructor `U` mistakenly referred to the no
longer existing `unsafeToInteger`.

Co-authored-by: Peter Lebbing <peter@digitalbrains.com>