Track usage in netlist
When producing sequential and concurrent code, there are restrictions
for how the same declared signal can be assigned which depend on the
HDL being targeted. In short, these are

    VHDL: A `signal` can be assigned continuously, or procedurally
    with non-blocking assignment. A `variable` can only be assigned
    procedurally with blocking assignment.

    Verilog: A `wire` can only be assigned continuously. A `reg`
    can only be assigned procedurally, but can be assigned both
    non-blocking and blocking.

When the backend produces HDL, it needs to know the type of a
declaration to render it. With the `WireOrReg` type, this was easy
for VHDL, but does not help differentiate between if a signal is
intended to be a VHDL signal or variable. A more general type is
now provided which provides the extra information to make this
choice:

```haskell
data Blocking = Blocking | NonBlocking
data Usage    = Cont | Proc Blocking
```

Since a declaration can accept more than one type of assignment,
the usage is not tracked with the declaration like `WireOrReg` was,
but tracked in the assignments and a usage map is created for every
component (tracking declared signals to their most restrictive use).
When producing netlist, the current use of a signal should be
checked to ensure that generated code only produces assignments
that are valid for the backend being targeted.