Store multiple `Usage` for black boxes
Depending on the current style of HDL being produced, we want to
be able to change the template of black boxes. However, this may
change the type of assignment to the result signal in the template.

Consider a template for an `fmap` implementation. In a concurrent
context we may choose to produce a generate statement, which uses
continuous assignment to the result variable. In a sequential
context we would instead use a for loop with procedural assignment.