Add ~ISSEQUENTIAL template tag
There is now a tag to test if the black box is being rendered in
a context which is concurrent or sequential. This allows the
template to provide a different implementation when a black box
can be implemented in both styles.

Somewhat annoyingly, templates can currently be rendered in the
wrong context with no warning until the HDL is given to a vendor
tool. It may be the case that many templates need to be wrapped
with something like

```
~IF ~ISSEQUENTIAL ~THEN ~ERR["no implementation possible"] ~ELSE ... ~FI
```

or vice-versa for templates that do not allow use in concurrent
contexts.