Render ~ARG[n] enclosed in parentheses
This makes sure there are no unexpected operator precedence issues
between operators in the blackbox and operators in the expression of the
~ARG.

This might cause a few superfluous parentheses somewhere.
But it fixes a lot of missing cases, like discovered in #2213.
And clash is smart enough to leave them out when the ~ARG is
simple enough, like a literal or a variable.

Fixes #2213

(cherry picked from commit 88ef335b3bfdbb1fbfd019e88c8c7c45763d1e40)