Add flag rendering to 'ClashOpts'
This adds a more structured way to define Clash command line flags to
`Clash.GHC.ClashFlags`. This comes with a couple of improvements over
the status quo:

  * Each boolean flag now defines two flags: `-fclash-foo` and
    `-fclash-no-foo`. This makes sure users can always unset a flag
    after setting it. This can be helpful in certain environments, and
    should help us once we'll support module-level flags.

  * Each flag backed by a `Maybe` can now be unset using `-fclash-no-foo`

Last but not least, `ClashOpts` can now be rendered to a set of flags.
This can help users of `clash-shake` to define a set of options using
`ClashOpts`, and then pass them to clash-the-binary.