Speed up, harmonize BitReduction
The cases for `reduceAnd` and `reduceOr` where the mask is non-zero can
also be sped up just like the cases with a zero mask.

`reduceAnd` and `reduceOr` always produce a `Bit`, but possibly with a
non-zero mask (i.e., an undefined value). `reduceXor` should also
produce such a value, which is different from throwing an `XException`
as it did.