Use `ClashEnv` in Netlist Generation
The netlist environment in the netlist monad now re-uses `ClashEnv`
instead of extracting a lot of it into `NetlistState`. At this
point it might be worth looking into using compact regions for the
`ClashEnv` since it may be passed directly to all main parts of the
compiler without being partially unpacked into some state or
environment type in a particular step.