DEC shouldn't care about the shape of proofs
Whether we have `Refl : True ~ True` or `SomeAxiom : (1 <=? 2) ~ True`
it doesn't matter, since when we normalize both sides we always end
up with a proof of `True ~ True`.

Since DEC only fires for applications where all the type arguments
are equal, we can deduce that all equality arguments witness the same
equality, hence we don't have to care about the shape of the proof.

Fixes #2289