New: Bundled clashConnector.tcl
We now bundle a Tcl script that parses `clash-manifest.json` files and
issues the correct Vivado Tcl commands to read the sources in
Non-Project Mode.

The Clash<->Tcl API for creating Xilinx IP will be added soon; first,
its features need to be extended.

The script can be obtained from `Clash.DataFiles.tclConnector` in
`clash-lib` or by invoking the tool `static-files`. The name of the tool
is generic so we can add different output products to it later and still
have the name make sense.