Tcl connector: implement API v1
A breaking change was made to the API. I realized the extra indirection
of defining a procedure was not actually needed in order to elegantly
pass in the argument specifying where to create the namespace.

Vivado tests in `clash-testsuite` now use the Tcl connector, and handle
a top component name that differs from the Haskell function that
specifies that component. Closes: #2264