Split `VarEnv` / `VarSet` / `InScopeSet` to modules
The `VarSet` and `InScopeSet` types exist for slightly different
reasons, they should live in their own modules which can be imported
qualified, and their functions given more consistent names like
`UniqMap` was earlier.

We can also remove `Clash.Core.VarEnv`, and expose `VarEnv` as a type
synonym from `Clash.Core.Var`. This is more consistent with other types
which are defined as synonyms, and makes it no longer possible for the
API of a `VarEnv` and `UniqMap` to fall out of sync.