CI: fix `nix-build` for newer kernels, clean up (#2339)
We need to install `awk` before we can use it in `effective_cpus.sh`.
`awk` is only used for newer kernels, like the one in Ubuntu 22.04. So
the previous code succeeded only on older runner hosts.

The shebang line for `effective_cpus.sh` is changed to pick `bash` from
the $PATH; it is pre-installed but not in /usr/bin.

An obsolete workaround for Circle CI and unneeded packages are removed.

It seems we only need to install `gawk` in the `before_script` (per
above). `nix-build` will automatically fetch the following packages
previously specified in the `before_script`:
findutils gnugrep gnused coreutils bash

These previously specified packages now seem unused: git ssh zstd.