Add `clockTicks` and `clockTicksEither`
Dual clock primitives need to decide when to sample/yield elements from
their respective signals. In order to make this decision, they need to
keep relative time. Subtle differences in this implementations can lead
to hard-to-debug errors (see #2346). This commit therefore introduces an
abstraction for this, making it easier to apply the logic consistently
and cleaning up primitives as a bonus.

Future commits will refactor existing primitives to use these functions.

Co-authored-by: Peter Lebbing <peter@digitalbrains.com>