Tcl connector: add verbosity, dry-run
`clash::verbosity` sets the logging level and the boolean
`clash::dryRun` enables printing of actions instead of performing them.
Log level 1, the default, is as verbose as the script was before. Log
lines are now prefixed by `Clash [n]:` where _n_ is the log level of the
message that follows.

Documentation will be added later.