Test suite: fix unintended test bench structure
The `RotateC` tests were written in a way that causes a functionally
identical copy of the `topEntity` to be included in the `testBench`,
instead of the usual case where the `testBench` uses a `topEntity` in
its own entity directory.

I can only think of one functional difference: such an included version
of the code will likely elide many enable lines as they are driven by
`enableGen`. With a separate entity directory, this transformation is
omitted because the `topEntity` has actual enable lines. Clearly if the
intent is to test the `~ISACTIVEENABLE` mechanism and similar
mechanisms, this is problematic, but this was not the case in these test
benches. However, it is clearly generally intended that the top entity
lives in its own entity dir, and the code surreptitiously does something
else than a cursory reading or general expectations suggest. In the
future, as some feature is added to Clash, it might even have more
consequences: the deviating structure might not be accounted for due to
it being unintended.