Record Integer and Natural constructors
While GHC might have dropped Integer and Natural literals, in
Clash it is still nice to have them around. However, Integer
and Natural are also no longer primitive types in GHC, but we
still want to give the Integer and Natural type to our Integer
and Natural literals.

So instead of recording the primitive types, we record the
algebraic types, i.e. the complete data type for Integer
and Natural, data constructors and all.