Tcl connector: Handle plain constraint files
When the Tcl connector finds a file named `*.sdc` or `*.xdc`, and it is
not handled by an accompanying Clash<->Tcl interface file, it will issue
a `read_xdc` for that file. Exception: for every top component (as
declared in `clash-manifest.json`), it will only load the `.sdc` file
with the same file name as the top component name for the actual top
entity in the design, not for subentities. Clash emits `create_clock`
statements in those files, and we should only load it for the actual top
entity.

Note that because all `.sdc` and `.xdc` files are loaded with an
unadorned `read_xdc <file>`, they should use globally unique identifiers
to match only the intended signals, entities and so on. That way, they
will not unintendedly match things outside the correct entities as the
statements are not scoped to any entity explicitly.