Don't quote bit literals with -fclash-force-undefined (#2378)
When rendering bit literals, we use the `char` function from the
prettyprinter in most cases. For undefined bits we were using the
`int` function, and wrapping in single quotes. This is not the same:
the `char` function prints a single Haskell `Char`, but does _not_
wrap that character in quotes.

Simply removing the quote marks solves the issue.

Fixes #2360.