GitLab CI: avoid recompilation, use matrix
Gitlab CI now supports `parallel:matrix` trigger jobs, so use that.

The tests that should get a complete build from the `build` job in
`test.yml` sometimes recompiled everything in GHC 9.2. The tests are
changed to simply invoke the prebuilt binary, so whatever goes wrong, it
will at least never recompile and burn CI time.

Now that we only invoke binaries, we can dispense with the cache and
`.ci/setup.sh` for doctests and unit tests.