Improve readability of Clash.Clocks.Deriving
No functional change, except that the `clocks` function is now annotated
with a `CLASH_OPAQUE` "pragma" rather than a mere `NOINLINE`.

By using TH quotations where possible and doing basically everything in
the Q monad instead of mixing it with pure constructors, the code is
much easier to grok.

Quotations and TH `Name`s don't interact very well.
`Clash.Clocks.Deriving` was renamed to `...Internal`, and by moving the
`Clocks` class to `Clash.Clocks.Internal`, we can use its names in the
quotation in `deriveClocksInstance` without resorting to TH `Name`s.
This means `Clash.Clocks` now contains the orphaned instances. This
could unfortunately not be fixed with a .hs-boot file because Template
Haskell doesn't play nice with that.

`Clash.Clocks.Internal` is not exported by `clash-prelude` so users will
not accidentally import the `Clocks` class without its instances and be
confused.

When building the Haddock, only emit instances for up to 3 clocks for
the `Clocks` class to reduce the amount of instances under the `Clock`
and `Signal` data types.