Improve readability of Clash.Clocks.Deriving
No functional change

By using TH quotations where possible and doing basically everything in
the Q monad instead of mixing it with pure constructors, the code is
much easier to grok.

Quotations and TH `Name`s don't interact very well. By moving the
`Clocks` class to `Clash.Clocks.Deriving`, we can use its names in the
quotation in `deriveClocksInstance` without resorting to TH `Name`s.
This means `Clash.Clocks` now contains the orphaned instances. This
could not be fixed with a .hs-boot file because Template Haskell doesn't
play nice with that.

`Clash.Clocks.Deriving` is no longer exported by `clash-prelude` so
users will not accidentally import the `Clocks` class without its
instances and be confused.

When building the Haddock, only emit instances for up to 3 clocks for
the `Clocks` class to reduce the amount of instances under the `Clock`
and `Signal` data types.